chore: split PairingModal into PairingHooks, PairingUIComponents, ShowCodeTab, EnterCodeTab
